requirements.txt
discord.py==2.2.2
yt-dlp==2023.3.4
youtube-search-python==1.6.6
PyNaCl==1.5.0
advocate==1.0.0
urllib3>=1.22,<2.0
spotipy==2.19.0
asyncio==3.4.3
